custom_nodes/comfyui-fal-splat/requirements.txt
fal-client>=0.5.0
requests>=2.28.0
Pillow>=9.0.0
numpy>=1.24.0
torch>=2.0.0
